requirements.txt
streamlit==1.32.0
pandas==2.1.4
mysql-connector-python==8.2.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.28
nba_api==1.3.1
